requirements.txt
minimalmodbus
psycopg2-binary